+++ src/race/race_manager.hpp
@@ -514,7 +514,10 @@ public:
      *  AI karts. */
     unsigned int getNumberOfKarts() const {return m_num_karts; }
     // ------------------------------------------------------------------------
-    unsigned int getNumberOfAIKarts() const {return m_ai_kart_list.size(); }
+    unsigned int getNumberOfAIKarts() const
+    {
+        return (unsigned int)m_ai_kart_list.size(); 
+    }   // getNumberOfAIKarts
     // ------------------------------------------------------------------------
     unsigned int getNumNonGhostKarts() const { return m_num_karts - m_num_ghost_karts; }
     // ------------------------------------------------------------------------
