+++ src/replay/replay_play.cpp
@@ -328,10 +328,12 @@ void ReplayPlay::loadFile(bool second_replay)
         return;
     }
 
-    Log::info("Replay", "Reading replay file '%s'.", getReplayFilename(replay_file_number).c_str());
+    Log::info("Replay", "Reading replay file '%s'.", 
+               getReplayFilename(replay_file_number).c_str());
 
     ReplayData &rd = m_replay_file_list[replay_index];
-    unsigned int num_kart = m_replay_file_list.at(replay_index).m_kart_list.size();
+    unsigned int num_kart = (unsigned int)m_replay_file_list.at(replay_index)
+                                                            .m_kart_list.size();
     unsigned int lines_to_skip = (rd.m_replay_version == 3) ? 7 : 10;
     lines_to_skip += (rd.m_replay_version == 3) ? num_kart : 2*num_kart;
 
@@ -358,13 +360,15 @@ void ReplayPlay::readKartData(FILE *fd, char *next_line, bool second_replay)
 {
     char s[1024];
 
-    int replay_index = second_replay ? m_second_replay_file : m_current_replay_file;
+    int replay_index = second_replay ? m_second_replay_file
+                                     : m_current_replay_file;
 
-    const unsigned int kart_num = m_ghost_karts.size();
+    const unsigned int kart_num = (unsigned int)m_ghost_karts.size();
     unsigned int first_loaded_f_num = 0;
 
     if (!second_replay && m_second_replay_enabled)
-        first_loaded_f_num = m_replay_file_list.at(m_second_replay_file).m_kart_list.size();
+        first_loaded_f_num = (unsigned int)m_replay_file_list.at(m_second_replay_file)
+                                                             .m_kart_list.size();
 
     ReplayData &rd = m_replay_file_list[replay_index];
     m_ghost_karts.push_back(std::make_shared<GhostKart>
